refactor(Jam): jsx to tsx
