+++ src/context/ServiceInfoContext.tsx
@@ -239,4 +239,11 @@ const useSessionConnectionError = () => {
   return context.connectionError
 }
 
-export { ServiceInfoContext, ServiceInfoProvider, useServiceInfo, useReloadServiceInfo, useSessionConnectionError }
+export {
+  ServiceInfoContext,
+  ServiceInfoProvider,
+  useServiceInfo,
+  useReloadServiceInfo,
+  useSessionConnectionError,
+  Schedule,
+}
